Extract and refactor filters into filters.py
This will help in implementing __in as a filter, refs #433
